src/internlm-math-plus/requirements.txt
aiosignal==1.3.1
annotated-types==0.6.0
antlr4-python3-runtime==4.7.2
anyio @ file:///croot/anyio_1706220167567/work
argon2-cffi @ file:///opt/conda/conda-bld/argon2-cffi_1645000214183/work
argon2-cffi-bindings @ file:///tmp/build/80754af9/argon2-cffi-bindings_1644553347904/work
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
attrs @ file:///croot/attrs_1695717823297/work
beautifulsoup4 @ file:///croot/beautifulsoup4-split_1681493039619/work
bleach @ file:///opt/conda/conda-bld/bleach_1641577558959/work
certifi==2024.2.2
cffi @ file:///croot/cffi_1700254295673/work
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
comm @ file:///croot/comm_1709322850197/work
cupy-cuda12x==12.1.0
debugpy @ file:///croot/debugpy_1690905042057/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
diskcache==5.6.3
entrypoints @ file:///tmp/build/80754af9/entrypoints_1649908313000/work
exceptiongroup @ file:///croot/exceptiongroup_1706031385326/work
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
fastapi==0.110.0
fastjsonschema @ file:///opt/conda/conda-bld/python-fastjsonschema_1661371079312/work
fastrlock==0.8.2
filelock==3.13.1
frozenlist==1.4.1
fsspec==2024.3.1
h11==0.14.0
httptools==0.6.1
huggingface-hub==0.22.0
idna @ file:///croot/idna_1666125576474/work
interegular==0.3.3
ipykernel @ file:///croot/ipykernel_1705933831282/work
ipython @ file:///croot/ipython_1704833016303/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
jedi @ file:///tmp/build/80754af9/jedi_1644315229345/work
Jinja2 @ file:///croot/jinja2_1706733616596/work
joblib==1.3.2
jsonschema @ file:///croot/jsonschema_1699041609003/work
jsonschema-specifications==2023.12.1
jupyter-events @ file:///croot/jupyter_events_1699282461638/work
jupyter_client @ file:///croot/jupyter_client_1676329080601/work
jupyter_core @ file:///croot/jupyter_core_1698937308754/work
jupyter_server @ file:///croot/jupyter_server_1699466442171/work
jupyter_server_terminals @ file:///croot/jupyter_server_terminals_1686870725608/work
jupyterlab-pygments @ file:///croot/jupyterlab_pygments_1700168593176/work
kiwisolver==1.4.5
lark==1.1.9
latex2sympy2==1.9.1
llvmlite==0.42.0
MarkupSafe @ file:///croot/markupsafe_1704205993651/work
matplotlib-inline @ file:///opt/conda/conda-bld/matplotlib-inline_1662014470464/work
mistune @ file:///opt/conda/conda-bld/mistune_1661496219659/work
mpmath==1.3.0
msgpack==1.0.8
nb-conda-kernels @ file:///home/builder/ci_310/nb_conda_kernels_1640815331881/work
nbclassic @ file:///croot/nbclassic_1699542793266/work
nbclient @ file:///croot/nbclient_1698934205032/work
nbconvert @ file:///croot/nbconvert_1699022732553/work
nbformat @ file:///croot/nbformat_1694616755618/work
nest-asyncio @ file:///croot/nest-asyncio_1708532673751/work
networkx==3.2.1
ninja==1.11.1.1
notebook @ file:///croot/notebook_1690984815942/work
notebook_shim @ file:///croot/notebook-shim_1699455894279/work
numba==0.59.1
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.7.0.84
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.19.3
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.1.105
outlines==0.0.37
overrides @ file:///croot/overrides_1699371140756/work
packaging @ file:///croot/packaging_1710807400464/work
pandocfilters @ file:///opt/conda/conda-bld/pandocfilters_1643405455980/work
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pillow==10.2.0
platformdirs @ file:///croot/platformdirs_1692205439124/work
prometheus-client @ file:///tmp/abs_d3zeliano1/croots/recipe/prometheus_client_1659455100375/work
prompt-toolkit @ file:///croot/prompt-toolkit_1704404351921/work
protobuf==5.26.0
psutil @ file:///opt/conda/conda-bld/psutil_1656431268089/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==2.6.4
pydantic_core==2.16.3
Pygments @ file:///croot/pygments_1684279966437/work
pynvml==11.5.0
pyparsing==3.1.2
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-dotenv==1.0.1
python-json-logger @ file:///croot/python-json-logger_1683823803357/work
PyYAML @ file:///croot/pyyaml_1698096049011/work
pyzmq @ file:///croot/pyzmq_1709318311506/work
ray==2.10.0
referencing @ file:///croot/referencing_1699012038513/work
regex==2023.12.25
requests==2.31.0
rfc3339-validator @ file:///croot/rfc3339-validator_1683077044675/work
rfc3986-validator @ file:///croot/rfc3986-validator_1683058983515/work
rpds-py @ file:///croot/rpds-py_1698945930462/work
safetensors==0.4.2
scipy==1.12.0
Send2Trash @ file:///croot/send2trash_1699371139552/work
sentencepiece==0.2.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio @ file:///croot/sniffio_1705431295498/work
soupsieve @ file:///croot/soupsieve_1696347547217/work
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
starlette==0.36.3
sympy==1.7.1
terminado @ file:///croot/terminado_1671751832461/work
tinycss2 @ file:///croot/tinycss2_1668168815555/work
tokenizers==0.15.2
torch==2.1.2
tornado @ file:///croot/tornado_1696936946304/work
tqdm==4.66.2
traitlets==5.14.2
transformers==4.39.1
triton==2.1.0
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
vllm @ https://github.com/vllm-project/vllm/releases/download/v0.3.3/vllm-0.3.3+cu118-cp310-cp310-manylinux1_x86_64.whl#sha256=9ff9cd9b21655fe57167a250a61cc53f60a7bf5b76506ce3c7d74d13ad8b07f6
watchfiles==0.21.0
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webencodings==0.5.1
websocket-client @ file:///home/builder/ci_310/websocket-client_1640795866898/work
websockets==12.0
xformers==0.0.23.post1
